refactor: player의 결과 계산 기능을 ResultCalculator 에서 player로 옮김
